requirements.txt
streamlit
requests
streamlit_lottie
Pillow